// car-part-front-micro/angular/src/app/core/api/v1/model/partFilterRequest.ts
/**
 * car-part-core openapi spec
 *
 * Contact: blabla@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PartFilterRequest { 
    carModel?: string;
    manufacturerName?: string;
    partName?: string;
    partManufacturerName?: string;
    startPrice?: number;
    endPrice?: number;
    page: number;
    pageElements: number;
}